Change type from int to float in is_in_intervall method

+++ exercise3.py
@@ -40,7 +40,7 @@ class Rectangle:
     def _is_idx_on_right_edge(self, i: int) -> bool:
         return i in [1, 3]
 
-    def is_in_interval(self, value: int, min_value: int, max_value: int) -> bool: # Task A
+    def is_in_interval(self, value: float, min_value: float, max_value: float) -> bool: # Task A
         return min_value <= value <= max_value
